requirements.txt
#
# This file is autogenerated by pip-compile with python 3.8
# To update, run:
#
#    pip-compile requirements.in
#
asgiref==3.4.1
    # via django
attrs==21.2.0
    # via jsonschema
authlib==0.15.4
    # via drf-oidc-auth
cachetools==4.2.2
    # via django-helusers
certifi==2021.5.30
    # via
    #   elasticsearch
    #   requests
    #   sentry-sdk
cffi==1.14.6
    # via cryptography
charset-normalizer==2.0.4
    # via requests
cryptography==3.4.8
    # via
    #   authlib
    #   drf-oidc-auth
    #   social-auth-core
defusedxml==0.7.1
    # via
    #   python3-openid
    #   social-auth-core
deprecation==2.1.0
    # via django-helusers
django==3.2.12
    # via
    #   -r requirements.in
    #   django-cors-headers
    #   django-etuovi
    #   django-helusers
    #   django-oikotie
    #   djangorestframework
    #   djangorestframework-simplejwt
    #   drf-oidc-auth
    #   drf-spectacular
django-cors-headers==3.7.0
    # via -r requirements.in
django-enumfields==2.1.1
    # via -r requirements.in
django-environ==0.4.5
    # via -r requirements.in
django_etuovi @ git+https://github.com/City-of-Helsinki/django-etuovi@v0.2.0
    # via -r requirements.in
django-helusers==0.7.0
    # via -r requirements.in
django_oikotie @ git+https://github.com/City-of-Helsinki/django-oikotie@v0.2.0
    # via -r requirements.in
django-pgcrypto-fields @ git+https://github.com/incuna/django-pgcrypto-fields@5b4b8eed3f9636ba3331f26e41a639fe4bee9247
    # via -r requirements.in
django-simple-history==3.0.0
    # via -r requirements.in
djangorestframework==3.12.4
    # via
    #   -r requirements.in
    #   djangorestframework-simplejwt
    #   drf-oidc-auth
    #   drf-spectacular
djangorestframework-simplejwt==4.7.2
    # via -r requirements.in
docopt==0.6.2
    # via num2words
drf-oidc-auth==2.0.0
    # via -r requirements.in
drf-spectacular==0.18.2
    # via -r requirements.in
ecdsa==0.17.0
    # via python-jose
elasticsearch==7.14.0
    # via
    #   -r requirements.in
    #   elasticsearch-dsl
elasticsearch-dsl==7.4.0
    # via -r requirements.in
hashids==1.3.1
    # via -r requirements.in
idna==3.2
    # via requests
inflection==0.5.1
    # via drf-spectacular
jsonschema==3.2.0
    # via drf-spectacular
lxml==4.6.5
    # via
    #   django-etuovi
    #   django-oikotie
    #   pikepdf
num2words==0.5.10
    # via -r requirements.in
oauthlib==3.1.1
    # via
    #   requests-oauthlib
    #   social-auth-core
packaging==21.0
    # via
    #   deprecation
    #   pikepdf
pikepdf==5.0.0
    # via -r requirements.in
pillow==9.0.1
    # via pikepdf
psycopg2==2.9.1
    # via -r requirements.in
pyasn1==0.4.8
    # via
    #   python-jose
    #   rsa
pycparser==2.20
    # via cffi
pyjwt==2.1.0
    # via
    #   djangorestframework-simplejwt
    #   social-auth-core
pyparsing==2.4.7
    # via packaging
pyrsistent==0.18.0
    # via jsonschema
python-dateutil==2.8.2
    # via elasticsearch-dsl
python-jose==3.3.0
    # via django-helusers
python-stdnum==1.17
    # via -r requirements.in
python3-openid==3.2.0
    # via social-auth-core
pytz==2021.1
    # via django
pyyaml==5.4.1
    # via drf-spectacular
requests==2.26.0
    # via
    #   django-helusers
    #   drf-oidc-auth
    #   requests-oauthlib
    #   social-auth-core
requests-oauthlib==1.3.0
    # via social-auth-core
rsa==4.7.2
    # via python-jose
sentry-sdk==1.3.1
    # via -r requirements.in
six==1.16.0
    # via
    #   ecdsa
    #   elasticsearch-dsl
    #   jsonschema
    #   python-dateutil
social-auth-app-django==5.0.0
    # via -r requirements.in
social-auth-core==4.1.0
    # via social-auth-app-django
sqlparse==0.4.2
    # via django
uritemplate==3.0.1
    # via drf-spectacular
urllib3==1.26.6
    # via
    #   elasticsearch
    #   requests
    #   sentry-sdk
whitenoise==5.3.0
    # via -r requirements.in

# The following packages are considered to be unsafe in a requirements file:
# setuptools
